+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 + Fixed "Anti Ghost Block" module from breaking blocks 
 + Fixed "Avoid" module not adding collision boxes to end portals
 + Added "Reset VL" module
++ Added "Nuker" module
 ```
 
 ---